+++ src/main/java/lv/venta/EATSystem/services/impl/EmployeeServiceImpl.java
@@ -30,7 +30,9 @@ public class EmployeeServiceImpl implements IEmployeeService{
 
 	@Override
 	public ArrayList<Employee> deleteEmployeeById(int id) {
+		if(employeeRepo.existsById(id)) {
 		employeeRepo.deleteByIdEmployee(id);
+		}
 		ArrayList<Employee> result = selectAllEmployees();
 		return result;
 	}
@@ -39,6 +41,7 @@ public class EmployeeServiceImpl implements IEmployeeService{
 	public Employee insertNewEmployee(String name, String surname, String position, Department department,
 			String email, boolean isManager) {
 		Employee result = new Employee(name, surname, position, department, email, isManager);
+		employeeRepo.save(result);
 		return result;
 	}
 
@@ -46,7 +49,13 @@ public class EmployeeServiceImpl implements IEmployeeService{
 	public Employee updateEmployeeById(int id, String name, String surname, String position, Department department,
 			String email, boolean isManager) throws Exception {
 		Employee result = employeeRepo.findByIdEmployee(id);
-		result = new Employee(name, surname, position, department, email, isManager);
+		result.setName(name);
+		result.setSurname(surname);
+		result.setPostion(position);
+		result.setDepartment(department);
+		result.setEmail(email);
+		result.setManager(isManager);
+		employeeRepo.save(result);
 		return result;
 	}
